+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.3.1] - 2026-08-25
+
 ### Fixed
 
 - **`htpolynet postsim` and `htpolynet analyze` no longer die in a directory